#571 WIP: Check the font size inherited from the console

+++ src/netxs/desktopio/directvt.hpp
@@ -921,7 +921,7 @@ namespace netxs::directvt
                     if (c1.img() != c2.img()) { meaning += sizeof(c1.img()); changes |= rastr; }
                     if (c1.egc() != c2.egc())
                     {
-                        cluster = c1.len();
+                        cluster = (byte)c1.len();
                         meaning += cluster + 1;
                         changes |= glyph;
                     }
